backend/requirements.txt
django==4.1.1
djangorestframework==3.13.1
django-cors-headers==3.13.0

djoser==2.1.0

# django-allauth==0.51.0
# dj-rest-auth==2.2.5

# drf-spectacular==0.23.1